pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qualytics-cli"
version = "0.4.0"
description = "Qualytics CLI - Command-line interface for the Qualytics data quality platform"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Qualytics", email = "devops@qualytics.ai"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["Qualytics", "Data Quality", "CLI", "Data Observability"]
dependencies = [
    "typer>=0.9.0",
    "requests>=2.28.0",
    "pyjwt>=2.6.0",
    "croniter>=1.3.0",
    "typing-extensions>=4.4.0",
    "rich>=13.0.0",
    "shellingham>=1.5.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "urllib3>=2.0.0",
    "openpyxl>=3.1.0",
    "fastmcp>=2.0.0",
]

[project.urls]
Homepage = "https://www.qualytics.ai/"
Repository = "https://github.com/Qualytics/qualytics-cli"
Documentation = "https://qualytics.github.io/userguide/"
Issues = "https://github.com/Qualytics/qualytics-cli/issues"
Changelog = "https://github.com/Qualytics/qualytics-cli/releases"

[project.scripts]
qualytics = "qualytics.qualytics:app"

[tool.hatch.build.targets.wheel]
packages = ["qualytics"]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]

[tool.coverage.run]
source = ["qualytics"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = ["E231", "E402", "E702", "E501", "F403", "F541", "E713"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["qualytics"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
